requirements.txt
# ─── Core Data & ML ───────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
optuna==3.6.1
joblib==1.4.2

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.5.0
psycopg2-binary==2.9.9

# ─── HTTP / API ────────────────────────────────────────────────────────────────
requests==2.32.3
python-dotenv==1.0.1

# ─── Model Storage ────────────────────────────────────────────────────────────
huggingface-hub==0.23.2

# ─── Dashboard ────────────────────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ─── Scheduler (local dev) ────────────────────────────────────────────────────
apscheduler==3.10.4

# ─── Notebooks ────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4
matplotlib==3.9.0
seaborn==0.13.2

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-mock==3.14.0
